Refactor: 코드 정리, 로그아웃 이후에도 유저 정보 남아있는 버그 수정
